requirements.txt
nltk~=3.7
torch~=1.11.0
torchvision
torchaudio
pyyaml==6.0
pika~=1.2.1
git+https://github.com/TartuNLP/fairseq.git@mtee-0.1.0
sentencepiece==0.1.96
pydantic~=1.9.1
huggingface-hub~=0.7.0
fastapi~=0.78.0
uvicorn~=0.17.6